Solucionado bloqueo vertical infinito y registro múltiple de puntuaciones

+++ src/packJuego/Pelota.java
@@ -5,8 +5,8 @@ import java.awt.Graphics;
 import packJuego.modificadores.Modificador;
 
 public class Pelota {
-	private int posx;
-	private int posy;
+	private Double posx;
+	private Double posy;
 	private double vecx;
 	private double vecy;
 
@@ -15,14 +15,13 @@ public class Pelota {
 	private int velocidad;
 
 	public Pelota(int pPosx, int pPosy, int pId) {
-		posx = pPosx;
-		posy = pPosy;
+		posx = (double) pPosx;
+		posy = (double) pPosy;
 		id = pId;
 		vecx = 1;
 		vecy = 0;
 		velocidad = DatosJuego.VELOCIDAD_INICIAL;
 		radio = DatosJuego.RADIO_PELOTA;
-
 
 	}
 
@@ -32,7 +31,7 @@ public class Pelota {
 	}
 
 	public void comprobarParedes() {
-		if ((posy <= DatosJuego.LIMITE_SUP && vecy  < 0) || (posy >= DatosJuego.ALTURA - radio * 2 && vecy  > 0)) {
+		if ((posy <= DatosJuego.LIMITE_SUP && vecy < 0) || (posy >= DatosJuego.ALTURA - radio * 2 && vecy > 0)) {
 			vecy *= -1;
 		}
 
@@ -45,7 +44,7 @@ public class Pelota {
 
 	public void pintar(Graphics g) {
 		g.setColor(DatosJuego.getColorPelota());
-		g.fillOval(posx, posy, radio * 2, radio * 2);
+		g.fillOval(posx.intValue(), posy.intValue(), radio * 2, radio * 2);
 	}
 
 	public void comprobarRaqueta(Raqueta pRaqueta) {
@@ -54,7 +53,7 @@ public class Pelota {
 			if (pRaqueta.getVelocidad() > 0) {
 				vecy += 1; // se suma el vector de la pelota con el vector (0,1)
 							// de la raqueta
-				normalizarVec(); 	// se normaliza el nuevo vector (se mantiene la
+				normalizarVec(); // se normaliza el nuevo vector (se mantiene la
 									// dirección, pero cambia el modulo a 1)
 			} else if (pRaqueta.getVelocidad() < 0) {
 				vecy -= 1;
@@ -62,13 +61,13 @@ public class Pelota {
 			}
 		}
 	}
-	
+
 	private void normalizarVec() {
 		double mag = Math.sqrt(vecx * vecx + vecy * vecy);
 		vecx = vecx / mag;
 		vecy = vecy / mag;
 	}
-	
+
 	private boolean colision(Raqueta pRaqueta) {
 		if (posy + 2 * radio >= pRaqueta.getPosy() && posy <= pRaqueta.getPosy() + DatosJuego.ALTURA_RAQUETA) {
 			if (pRaqueta.getPosx() == DatosJuego.LIMITE_IZQ) {
@@ -95,24 +94,26 @@ public class Pelota {
 	public void aumentarVelocidad() {
 		velocidad++;
 	}
-	
+
 	public int getPosy() {
-		return posy;
+		return posy.intValue();
 	}
 
 	public void comprobarModificadores() {
-		Modificador m = Juego.getJuego().buscarModificador(posx,posy);
+		Modificador m = Juego.getJuego().buscarModificador(posx.intValue(), posy.intValue());
 		if (m != null) {
 			System.out.println(m.getClass().getSimpleName());
-			switch(m.getClass().getSimpleName()) {
+			switch (m.getClass().getSimpleName()) {
 			case "Acelerador":
 				System.out.println("Acelerador activado");
-				velocidad*=2;
+				velocidad *= 2;
 				break;
 			case "Freno":
 				System.out.println("Freno activado");
-				if (vecx < 0) Juego.getJuego().ralentizarJ2();
-				else Juego.getJuego().ralentizarJ1();
+				if (vecx < 0)
+					Juego.getJuego().ralentizarJ2();
+				else
+					Juego.getJuego().ralentizarJ1();
 				break;
 			case "Duplicador":
 				System.out.println("Duplicador activado");
@@ -120,6 +121,5 @@ public class Pelota {
 				break;
 			}
 		}
-		
 	}
 }
